Accept and pass string in setUserId
While testing on android I noticed that the `setUserId` method in dart
accepts a string, but when the `userId` value is sent down to the swift
and kotlin files the logic there checks to see if they are integers and
if they are it coerces them into strings and then sends the strings to
raygun. If they are not integers it sets an anonymous user.

This changes the dart method so it accepts a dynamic type, checks to
make sure that `userId` is a `String` or an `int` and coerces an `int`
into a string to send to the kotlin and swift classes.